requirements.txt
accelerate==0.29.3
basicsr==1.4.2
diffusers==0.27.2
einops==0.3.0
gfpgan==1.3.8
insightface==0.7.3
invisible-watermark==0.1.5
kornia==0.6.0
matplotlib==3.7.5
omegaconf==2.1.1
onnxruntime-gpu==1.16.3
opencv-python==4.8.1.78
pandas==1.5.3
PyGObject==3.40.1
pytorch-lightning==1.4.2
realesrgan==0.3.0
safetensors==0.4.1
scikit-image==0.20.0
scipy==1.9.1
tensorboard==2.12.0
timm==0.6.12
torchmetrics==0.6.0
torchsde==0.2.6
tqdm==4.65.0
transformers==4.40.0
torchdiffeq==0.2.3
ftfy==6.2.0